refactor : service 단 통일로 인한 분기
